pyproject.toml
[project]
name = "projeto_01"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "gabrielimarta", email = "gabrielirocha2730@gmail.com"},
    {name = "", email = ""},
]
dependencies = [
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
